Place libraries behind object files to ensure symbols are known
Unless --as-needed is used as a linker argument, any unused symbol in
a library will be ignored and forgotten by the time we hit the object
files. Some architectures/linkers use `--as-needed` by default and
thus don't hit this issue, but in general, it's better to put the
libraries last (or be explicit about `--as-needed` in the Makefile).

And yes, in theory, the libraries also have to be in the correct order
if they depend on each other, but that's a whole other issue...

See also https://stackoverflow.com/q/45135