Replace Makefiles with CMakeLists.txt
This commit introduces a single top-level CMakeLists.txt to create the
necessary Makefiles (or other files, see CMake's -G option) for the
flux-compose project.

The [CMake] project is a "family of tools to build, test and package
software". While this sounds great, the real kicker is the ability to
create `compile_commands.json` files, which contain the compiler calls
with the necessary include paths and other arguments. This enables us
to use an LSP for the `flux-compose` project, such as `lsp-mode` or
`eglot`.

If you're not too familiar with CMake, a `bootstrap.sh` is provided to
create an initial build (called `build`) directory for you, but you
can also use an alternative build directory, e.g.

    ./bootstrap.sh Debug

will create the `Debug` directory instead. `boostrap.sh` is ONLY meant
to bring you up to speed and can be a starting point for further CMake
ventures, but does not replace proper CMake handling. Please don't add
too many features there.

[CMake]: https://cmake.org